Fix exports and move help in place
