+++ daemon/src/main/java/org/apache/maven/cli/CommonsCliDaemonMavenOptions.java
@@ -18,18 +18,31 @@
  */
 package org.apache.maven.cli;
 
+import java.io.ByteArrayOutputStream;
+import java.io.PrintStream;
+import java.io.PrintWriter;
+import java.nio.charset.StandardCharsets;
 import java.util.Collection;
 import java.util.List;
 import java.util.ListIterator;
+import java.util.Locale;
 import java.util.Map;
+import java.util.Set;
+import java.util.StringTokenizer;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 import org.apache.commons.cli.CommandLine;
+import org.apache.commons.cli.HelpFormatter;
 import org.apache.commons.cli.Option;
 import org.apache.commons.cli.Options;
 import org.apache.commons.cli.ParseException;
 import org.apache.maven.cling.invoker.mvn.CommonsCliMavenOptions;
+import org.apache.maven.jline.MessageUtils;
 import org.codehaus.plexus.interpolation.BasicInterpolator;
 import org.codehaus.plexus.interpolation.InterpolationException;
+import org.mvndaemon.mvnd.common.Environment;
+import org.mvndaemon.mvnd.common.OptionType;
 
 import static org.apache.maven.cling.invoker.Utils.createInterpolator;
 
@@ -81,6 +94,14 @@ public class CommonsCliDaemonMavenOptions extends CommonsCliMavenOptions impleme
     protected static class CLIManager extends CommonsCliMavenOptions.CLIManager {
         public static final String RAW_STREAMS = "ras";
 
+        private static final Pattern HTML_TAGS_PATTERN = Pattern.compile("<[^>]*>");
+        private static final Pattern COLUMNS_DETECTOR_PATTERN = Pattern.compile("^[ ]+[^s]");
+        private static final Pattern WS_PATTERN = Pattern.compile("\\s+");
+
+        static String toPlainText(String javadocText) {
+            return HTML_TAGS_PATTERN.matcher(javadocText).replaceAll("");
+        }
+
         @Override
         protected void prepareOptions(Options options) {
             super.prepareOptions(options);
@@ -88,6 +109,150 @@ public class CommonsCliDaemonMavenOptions extends CommonsCliMavenOptions impleme
                     .longOpt("raw-streams")
                     .desc("Use raw-streams for daemon communication")
                     .build());
+        }
+
+        @Override
+        public void displayHelp(String command, PrintWriter pw) {
+            ByteArrayOutputStream baos = new ByteArrayOutputStream();
+            try (PrintWriter pww = new PrintWriter(new PrintStream(baos), true, StandardCharsets.UTF_8)) {
+                super.displayHelp(command, pww);
+            }
+            final String mvnHelp = baos.toString(StandardCharsets.UTF_8);
+            final Matcher m = COLUMNS_DETECTOR_PATTERN.matcher(mvnHelp);
+            final String indent = m.find() ? m.group() : "                                        ";
+
+            final int terminalWidth = getTerminalWidth();
+            final String lineSeparator = System.lineSeparator();
+            final StringBuilder help =
+                    new StringBuilder(mvnHelp).append(lineSeparator).append("mvnd specific options:");
+
+            Environment.documentedEntries().forEach(entry -> {
+                final Environment env = entry.getEntry();
+                help.append(lineSeparator);
+                int indentPos = help.length() + indent.length();
+                int lineEnd = help.length() + terminalWidth;
+                spaces(help, HelpFormatter.DEFAULT_LEFT_PAD);
+                final String property = env.getProperty();
+                if (property != null) {
+                    help.append("-D").append(property);
+                    if (env.getType() != OptionType.VOID) {
+                        help.append("=<")
+                                .append(env.getType().name().toLowerCase(Locale.ROOT))
+                                .append('>');
+                    }
+                }
+
+                final Set<String> opts = env.getOptions();
+                if (!opts.isEmpty()) {
+                    if (property != null) {
+                        help.append(';');
+                    }
+                    boolean first = true;
+                    for (String opt : opts) {
+                        if (first) {
+                            first = false;
+                        } else {
+                            help.append(',');
+                        }
+                        help.append(opt);
+                    }
+                    if (env.getType() != OptionType.VOID) {
+                        help.append(" <")
+                                .append(env.getType().name().toLowerCase(Locale.ROOT))
+                                .append('>');
+                    }
+                }
+                help.append(' ');
+
+                spaces(help, indentPos - help.length());
+                wrap(help, toPlainText(entry.getJavaDoc()), terminalWidth, lineEnd, indent);
+
+                if (env.isDocumentedAsDiscriminating()) {
+                    indentedLine(help, terminalWidth, "This is a discriminating start parameter.", indent);
+                }
+                if (env.getDefault() != null) {
+                    indentedLine(help, terminalWidth, "Default: " + env.getDefault(), indent);
+                }
+                if (env.getEnvironmentVariable() != null) {
+                    indentedLine(help, terminalWidth, "Env. variable: " + env.getEnvironmentVariable(), indent);
+                }
+            });
+
+            help.append(lineSeparator).append(lineSeparator).append("mvnd value types:");
+
+            OptionType.documentedEntries().forEach(entry -> {
+                final OptionType type = entry.getEntry();
+                help.append(lineSeparator);
+                int indentPos = help.length() + indent.length();
+                int lineEnd = help.length() + terminalWidth;
+                spaces(help, HelpFormatter.DEFAULT_LEFT_PAD);
+                help.append(type.name().toLowerCase(Locale.ROOT));
+                spaces(help, indentPos - help.length());
+                wrap(help, toPlainText(entry.getJavaDoc()), terminalWidth, lineEnd, indent);
+            });
+
+            pw.print(help);
+            pw.flush();
+        }
+
+        private static int getTerminalWidth() {
+            int terminalWidth = MessageUtils.getTerminalWidth();
+            if (terminalWidth <= 0) {
+                terminalWidth = HelpFormatter.DEFAULT_WIDTH;
+            }
+            return terminalWidth;
+        }
+
+        private static void indentedLine(StringBuilder stringBuilder, int terminalWidth, String text, String indent) {
+            final int lineEnd = stringBuilder.length() + terminalWidth;
+            stringBuilder.append(System.lineSeparator()).append(indent);
+            wrap(stringBuilder, text, terminalWidth, lineEnd, indent);
+        }
+
+        /**
+         * Word-wrap the given {@code text} to the given {@link StringBuilder}
+         *
+         * @param stringBuilder the {@link StringBuilder} to append to
+         * @param text          the text to wrap and append
+         * @param lineLength    the preferred line length
+         * @param nextLineEnd   the length of the {@code stringBuilder} at which the current line should end
+         * @param indent        the indentation string
+         */
+        static void wrap(StringBuilder stringBuilder, String text, int lineLength, int nextLineEnd, String indent) {
+            final StringTokenizer st = new StringTokenizer(text, " \t\n\r", true);
+            String lastWs = null;
+            while (st.hasMoreTokens()) {
+                final String token = st.nextToken();
+                if (WS_PATTERN.matcher(token).matches()) {
+                    lastWs = token;
+                } else {
+                    if (stringBuilder.length() + token.length() + (lastWs != null ? lastWs.length() : 0)
+                            < nextLineEnd) {
+                        if (lastWs != null) {
+                            stringBuilder.append(lastWs);
+                        }
+                        stringBuilder.append(token);
+
+                    } else {
+                        nextLineEnd = stringBuilder.length() + lineLength;
+                        stringBuilder
+                                .append(System.lineSeparator())
+                                .append(indent)
+                                .append(token);
+                    }
+                    lastWs = null;
+                }
+            }
+        }
+
+        /**
+         * Append {@code count} spaces to the given {@code stringBuilder}
+         *
+         * @param  stringBuilder the {@link StringBuilder} to append to
+         * @param  count         the number of spaces to append
+         */
+        static void spaces(StringBuilder stringBuilder, int count) {
+            stringBuilder.append(" ".repeat(Math.max(0, count)));
         }
     }
 }
